Remove .psm1 from helpfile name
The file being created, Module.psm1-Help.xml is not being recognized by
the shell because of the .psm1.  I have removed that from the
Get-HelpFileName cmdlet.  This will not update any existing metadata,
but it will fix new ones.  I think this should work fine with the dll
modules, because then it needs to go Module.dll-Help.xml.