Removing the auto increment of help version
The help version is auto incremented when PlatyPS is run. This is done for
creating the cab files, so they have a higher version. The issue with this
approach is that the changed module.md file is not checked in, hence the
next time we run PlatyPS it will generate the same version as previous
run.

Fixed test that verfied auto increment of version.