Improve Example output formatting
The current formatting does not tag the code block as PowerShell. The remarks attribute sometimes is the formatted output of the example which are not within the code block. If there is a new line between the command example and the output of the command in the ".EXAMPLE", the output is considered remarks and will be unformatted by the current implementation. This provides, subjectively, a more esthetic output.

Here is an example of the output: https://github.com/roberttoups/IPv4Toolbox/blob/master/IPv4Toolbox/0.1.0/Docs/Invoke-IPv4ListSort.md