Change New-MarkdownCommandHelp dramatically.
Removed optional alpha sorting for parameters, this is always the case.
Change to use fulltypes for parameters by default.
Change to accept CommandInfo and PSModuleInfo objects.
A transform attribute will convert strings to objects rather than the transformer.
Add some overloads to the transformer to take those objects.
Files are grouped by module name, and placed in their own directory.
If they're not associated with a module (like a script), then they go into the 'outputfolder'. Fix a bug in the markdown reader to include the first input or output if it doesn't have a description.
Added a bunch of tests for new-markdowncommandhelp.