Change New-MarkdownCommandHelp dramatically. (#636)
* cp

* Change how we gather commands and modules for creating new markdown help.

* Add another overload for base module file metadata and use it.

* add transform from string to cmdlet and module.

update some tests, still some failures.

* Only write progress if there are more than 10 commands.

Skip tests that use the old modulefilepath parameter.

* a couple new overloads for base metadata for module files.

don't emit progress with fewer than 10 commands.
save off the module guids for use in the module file writer.

* Change New-MarkdownCommandHelp dramatically.

Removed optional alpha sorting for parameters, this is always the case.
Change to use fulltypes for parameters by default.
Change to accept CommandInfo and PSModuleInfo objects.
A transform attribute will convert strings to objects rather than the transformer.
Add some overloads to the transformer to take those objects.
Files are grouped by module name, and placed in their own directory.
If they're not associated with a module (like a script), then they go into the 'outputfolder'. Fix a bug in the markdown reader to include the first input or output if it doesn't have a description.
Added a bunch of tests for new-markdowncommandhelp.