Metadata is now a string/object dictionary,
since the value may be an array.
Create a serializer for that dictionary to support in flow style (array elements on a single line).
Add the dlls to the nested module list (as these need to be loaded before the module assembly is loaded due to the flow style change.
Explicitly sort the metadata before serializing it.
Fix parser bug for the last ast.
Change the object model for the module file due to requirement to handle multiple command groups (supporting exchange).
Add diagnostics for module file.