Refactor a number of files.
Change metadata dictionary to be <string,object> rather than <string,string>.
Add handling of HelpUri and HelpInfoUri to new-markdowncommandhelp (transform settings does not preserve these).
Update metadata serializer to handle arrays in flow format.
Add a reasonable implementation of ToString for DiagnosticMessages to improve formatting.
Be sure to include aliases in commandhelp.
Add support for metadata as a paragraph in addition to normal handling.
Add a number of diagnostic messages throughout.
Change alias boilerplate to appear only if the ALIAS header was missing. This means that the first conversion from V1 -> V2 will include boilerplate, but V2 -> V2 will not.
Don't add empty lines for the description of ModuleCommands, only if description is present.
Add a number of format entries for DiagnosticMessage, Iditify, Diagnostics (now a custom format).
Add Update-MarkdownModuleFile, Measure-PlatyPSMarkdown (identify what kind of markdown we have).
Add a bunch of tests