Additional updates for OPS release (#664)
Fix issue with module parsing when schema was not present.
Fix up tests for same and add exchange module file (for tests against multiple command groups.
Refactor a number of files.
Change metadata dictionary to be <string,object> rather than <string,string>.
Add handling of HelpUri and HelpInfoUri to new-markdowncommandhelp (transform settings does not preserve these).
Update metadata serializer to handle arrays in flow format.
Add a reasonable implementation of ToString for DiagnosticMessages to improve formatting.
Be sure to include aliases in commandhelp.
Add support for metadata as a paragraph in addition to normal handling.
Add a number of diagnostic messages throughout.
Change alias boilerplate to appear only if the ALIAS header was missing. This means that the first conversion from V1 -> V2 will include boilerplate, but V2 -> V2 will not.
Don't add empty lines for the description of ModuleCommands, only if description is present.
Add a number of format entries for DiagnosticMessage, Iditify, Diagnostics (now a custom format).
Add Update-MarkdownModuleFile, Measure-PlatyPSMarkdown (identify what kind of markdown we have).
* Add metadata during import of MAML file.
* Change parameter name to from OutputDirectory to OutputFolder for Export-MamlCommandHelp.
Be sure to save command help notes as AlertSet.
Be sure to create metadata when importing MAML file.
* fix issue in Measure-PlatyPSMarkdown.
It was not identifying a V2 module file. Add a test for same.
* Set diagnostics and ParameterNames as skipped property names for Compare-CommandHelp.
* updates to helpinfouri and helpuri. Be sure that they go in the right file.
* use CurrentCulture rather than CurrentUICulture
* Be sure to default metadata Locale to en-US.
* Skip implicit loading of module in Windows PowerShell.
* When testing, be sure to execute the right powershell.
* Fixes to build script to ensure testing on Windows PowerShell.
* Rewrite of tests to support running on Windows PowerShell.