Improve error message for Import-YamlModuleFile if the file is not a module file.
Be sure to set AliasHeaderFound to true to avoid emitting the boilerplate for alias when updating command help.
Remove extraneous using statements.
When importing yaml command help set the notes to an empty string if it is null.
Fix tests for Export-MarkdownCommandHelp.
Add additional tests for HelpUri to ensure it is not in the Update-MarkdownModule file output.
Check to be sure that HelpInfoUri _is_ present.